Threading > Success
+ What I thought was a race condition was in fact an issue with my scan
completion callback lambda, which captured the visualization parameters
by reference. Since the scanning operation is asynchronous, this meant
that the parameters went out of scope. Capturing this particular object
by value fixed the problem.
+ In order to be able to scan multiple times during a single session, I
removed the use of smart pointers for the QThread and worker in the
DriveScanner. I'll have to revisit this...