Refactor > Scene Assets
+ This is still a work in progress, but I want the canvas to only know
about "Scene Assets," instead of knowing about each individual item in
the scene (such as the treemap, the grid, et cetera). I think I'm fairly
close to getting this to work, but debugging this is going to take some
time. As such, all the new stuff is currently commented out. Everything
should still be functional.