Node Selection via AABB
+ Refined the selection algorithm a bit so that it only ends up
considering nodes whose bounding box AND block box have been
intersected; this prevents the problem described in the last commit.
* While the algorithm now seems to work a bit better, it appears that
getting too close to certain nodes still causes issues. My current
theory is that the camera might be within the bounding box of an
ancestor and that this screws up the bounding box intersection testing.
* On the plus side, node selection appears to be a lot quicker from my
brief tests so far.