Edge Cases
+ Some folders, like "C:\Documents and Settings," are pretty special,
and they need a tender touch. That is, some folders are special within
the operating system, while others exist purely as junctions for legacy
support reasons. In trying to figure out if a particular path is valid,
some boost::filesystem operations will attempt to create an iterator to
the file in question, while other operations forego this need. By
checking to see if a folder is a symlink before trying to see if it's
empty we can avoid any dangerous heavy lifting, as it were, and properly
handle the "C:\Documents and Settings" case. Since we're skipping
junctions, shortcuts, and symbolic links, there's never a need to check
the file size before checking for the aforementioned conditions.