File Scanning Bug Fix and Refactor
+ I wasn't including the regular files in the top level directory in the
scan. This is partially why the number of total bytes processed never
appeared to add up to anywhere close to 100%. The other issue is that
Windows hides a few critical files in the C:\ directory as well. These
files include paging and hibernation system files, which can be many
gigabytes in size. These files are also somewhat special in that Windows
does not grant access to them, and I haven't found a way yet to get
their size. So far any API queries to do so throw exceptions.
+ Switched to using `std::experimental::filesystem` instead of
`boost::filesystem`. This means that I no longer need to link in either
`boost::filesystem` or `boost::system`.
+ Refactored some of the scanning code to make it a bit less nested and
slightly more readable.
+ Now using the correct slash in the file path readout; `\` instead of
`/`.
+ Now imposing a limit of 4 scanning threads.